修复告警

+++ services/media_engine/plugins/ffmpeg_adapter/demuxer/ffmpeg_format_helper.cpp
@@ -464,7 +464,7 @@ static bool ConvertHexStrToBuffer(const std::string hexStr, std::vector<uint8_t>
     if (hexStr.empty() || hexStr.size() % VALUE_2 != 0) {
         return false;
     }
-    int len = hexStr.size() / VALUE_2;
+    int len = static_cast<int>(hexStr.size()) / VALUE_2;
     buffer.resize(len);
     for (int i = 0; i < len; i++) {
         int8_t high = HexCharToValue(hexStr[i * VALUE_2]);
